Another reorganization and simplification
