--- src/types/contexts/index.ts
@@ -1,3 +1,0 @@
-export * from './login'
-export * from './session'
-export * from './transact'
