--- src/types/contexts/session.ts
@@ -1,12 +1,0 @@
-import {APIClient} from '@greymass/eosio'
-
-export interface SessionContextOptions {
-    client: APIClient
-}
-
-export class SessionContext {
-    client: APIClient
-    constructor(options: SessionContextOptions) {
-        this.client = options.client
-    }
-}
