--- src/types/hooks/index.ts
@@ -1,2 +1,0 @@
-export * from './login'
-export * from './transact'
