+++ src/types/index.ts
@@ -1,7 +1,6 @@
-export * from './contexts'
-export * from './hooks'
-export * from './plugins'
-
+export * from './hook'
 export * from './kit'
+export * from './login'
 export * from './session'
 export * from './transact'
+export * from './wallet'
