+++ src/types/kit.ts
@@ -2,12 +2,9 @@ import {Checksum256Type, NameType, PermissionLevelType} from '@greymass/eosio'
 
 import {Session} from '../session'
 import {ChainDefinitionType, Fetch} from '../types'
-
-import {LoginPlugin, TransactPlugin, WalletPlugin} from './plugins'
-
-export abstract class AbstractSessionKit {
-    abstract login(options?: LoginOptions): Promise<Session>
-}
+import {LoginPlugin} from './login'
+import {TransactPlugin} from './transact'
+import {WalletPlugin} from './wallet'
 
 export interface LoginOptions {
     chain?: Checksum256Type
