--- src/types/plugins/index.ts
@@ -1,3 +1,0 @@
-export * from './login'
-export * from './transact'
-export * from './wallet'
