--- src/types/plugins/login.ts
@@ -1,9 +1,0 @@
-import {LoginContext} from '../contexts'
-
-export interface LoginPlugin {
-    register: (context: LoginContext) => void
-}
-
-export abstract class AbstractLoginPlugin implements LoginPlugin {
-    public abstract register(context: LoginContext): void
-}
