+++ src/types/session.ts
@@ -1,16 +1,12 @@
 import {APIClient, PermissionLevelType} from '@greymass/eosio'
 
-import {TransactArgs, TransactOptions, TransactResult} from './transact'
-import {AbstractTransactPlugin, WalletPlugin} from './plugins'
 import {ChainDefinitionType} from '../types'
+import {AbstractTransactPlugin} from './transact'
+import {WalletPlugin} from './wallet'
 
-export abstract class AbstractSession {
-    /**
-     * Transact using this session. See [[Session.transact]].
-     */
-    abstract transact(args: TransactArgs, options?: TransactOptions): Promise<TransactResult>
-}
-
+/**
+ * Options for creating a new instance of a [[Session]].
+ */
 export interface SessionOptions {
     chain: ChainDefinitionType
     client?: APIClient
